requirements.txt
requests
pillow
matplotlib
streamlit
numpy
pandas

